requirements.txt

--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.8.1+cpu
torchaudio==0.8.1
torchvision==0.9.1+cpu
numpy==1.20.0
pandas
streamlit
Pillow==8.2.0